@@ -5,7 +5,7 @@ classifiers =
     Programming Language :: Python :: 3
 
 [options]
-python_requires = >=3.6
+python_requires = >=3.7
 packages =
     bpython
     bpython.curtsiesfrontend
@@ -16,12 +16,12 @@ packages =
 install_requires =
     backports.cached-property; python_version < "3.8"
     curtsies >=0.3.5
-    dataclasses; python_version < "3.7"
     cwcwidth
     greenlet
     pygments
     pyxdg
     requests
+    typing-extensions
 
 [options.extras_require]
 clipboard = pyperclip
