@@ -343,12 +343,13 @@ class ImportCompletion(BaseCompletionType):
         return after_last_dot(word)
 
 
+def _safe_glob(pathname: str) -> Iterator[str]:
+    return glob.iglob(glob.escape(pathname) + "*")
+
+
 class FilenameCompletion(BaseCompletionType):
     def __init__(self, mode: AutocompleteModes = AutocompleteModes.SIMPLE):
         super().__init__(False, mode)
-
-    def safe_glob(self, pathname: str) -> Iterator[str]:
-        return glob.iglob(glob.escape(pathname) + "*")
 
     def matches(
         self, cursor_offset: int, line: str, **kwargs: Any
@@ -359,7 +360,7 @@ class FilenameCompletion(BaseCompletionType):
         matches = set()
         username = cs.word.split(os.path.sep, 1)[0]
         user_dir = os.path.expanduser(username)
-        for filename in self.safe_glob(os.path.expanduser(cs.word)):
+        for filename in _safe_glob(os.path.expanduser(cs.word)):
             if os.path.isdir(filename):
                 filename += os.path.sep
             if cs.word.startswith("~"):
@@ -371,7 +372,6 @@ class FilenameCompletion(BaseCompletionType):
         return lineparts.current_string(cursor_offset, line)
 
     def format(self, filename: str) -> str:
-        filename.rstrip(os.sep).rsplit(os.sep)[-1]
         if os.sep in filename[:-1]:
             return filename[filename.rindex(os.sep, 0, -1) + 1 :]
         else:
@@ -570,15 +570,15 @@ class ParameterNameCompletion(BaseCompletionType):
         r = self.locate(cursor_offset, line)
         if r is None:
             return None
-        if argspec:
-            matches = {
-                f"{name}="
-                for name in argspec[1][0]
-                if isinstance(name, str) and name.startswith(r.word)
-            }
-            matches.update(
-                name + "=" for name in argspec[1][4] if name.startswith(r.word)
-            )
+
+        matches = {
+            f"{name}="
+            for name in argspec[1][0]
+            if isinstance(name, str) and name.startswith(r.word)
+        }
+        matches.update(
+            name + "=" for name in argspec[1][4] if name.startswith(r.word)
+        )
         return matches if matches else None
 
     def locate(self, cursor_offset: int, line: str) -> Optional[LinePart]:
