@@ -21,6 +21,10 @@
 # THE SOFTWARE.
 
 import pydoc
+from types import TracebackType
+from typing import Optional, Type
+from typing_extensions import Literal
+
 from .. import _internal
 
 
@@ -29,8 +33,14 @@ class NopPydocPager:
         self._orig_pager = pydoc.pager
         pydoc.pager = self
 
-    def __exit__(self, *args):
+    def __exit__(
+        self,
+        exc_type: Optional[Type[BaseException]],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional[TracebackType],
+    ) -> Literal[False]:
         pydoc.pager = self._orig_pager
+        return False
 
     def __call__(self, text):
         return None
