@@ -26,11 +26,12 @@ import keyword
 import pydoc
 import re
 from collections import namedtuple
+from typing import Any, Optional, Type
+from types import MemberDescriptorType, TracebackType
+from typing_extensions import Literal
 
 from pygments.token import Token
 from pygments.lexers import Python3Lexer
-from typing import Any
-from types import MemberDescriptorType
 
 from .lazyre import LazyReCompile
 
@@ -94,7 +95,12 @@ class AttrCleaner:
         self.attribs = (__getattribute__, __getattr__)
         # /Dark magic
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
+    def __exit__(
+        self,
+        exc_type: Optional[Type[BaseException]],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional[TracebackType],
+    ) -> Literal[False]:
         """Restore an object's magic methods."""
         type_ = type(self.obj)
         __getattribute__, __getattr__ = self.attribs
@@ -104,6 +110,7 @@ class AttrCleaner:
         if __getattr__ is not None:
             setattr(type_, "__getattr__", __getattr__)
         # /Dark magic
+        return False
 
 
 class _Repr:
@@ -266,7 +273,7 @@ def getfuncprops(func, f,cls=None):
     return fprops
 
 
-def is_eval_safe_name(string):
+def is_eval_safe_name(string: str) -> bool:
     return all(
         part.isidentifier() and not keyword.iskeyword(part)
         for part in string.split(".")
@@ -321,7 +328,6 @@ def get_argspec_from_signature(f,cls=None):
 
     if not annotations:
         annotations = None
-
     return [
         args,
         varargs,
@@ -336,7 +342,7 @@ def get_argspec_from_signature(f,cls=None):
 get_encoding_line_re = LazyReCompile(r"^.*coding[:=]\s*([-\w.]+).*$")
 
 
-def get_encoding(obj):
+def get_encoding(obj) -> str:
     """Try to obtain encoding information of the source of an object."""
     for line in inspect.findsource(obj)[0][:2]:
         m = get_encoding_line_re.search(line)
@@ -356,7 +362,7 @@ def get_encoding_file(fname: str) -> str:
     return "utf8"
 
 
-def getattr_safe(obj: Any, name: str):
+def getattr_safe(obj: Any, name: str) -> Any:
     """side effect free getattr (calls getattr_static)."""
     result = inspect.getattr_static(obj, name)
     # Slots are a MemberDescriptorType
@@ -371,8 +377,3 @@ def hasattr_safe(obj: Any, name: str) -> bool:
         return True
     except AttributeError:
         return False
-
-
-def get_source_unicode(obj):
-    """Returns a decoded source of object"""
-    return inspect.getsource(obj)
